fix linux workflow

+++ include/hpp_proto/json_serializer.hpp
@@ -729,7 +729,7 @@ constexpr glz::opts get_glz_opts() {
 
 template <uint8_t width = 3>
 constexpr auto indent_level =
-    glz_opts_t<glz::opts{.prettify = static_cast<uint8_t>(width > 0), .indentation_width = width}>{};
+    glz_opts_t<glz::opts{.prettify = (width > 0), .indentation_width = width}>{};
 
 struct [[nodiscard]] json_status final {
   glz::error_ctx ctx;
